refactor(admin-ui): align RainbowButton imports
